requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
email-validator
python-multipart
passlib
pyyaml